# frontend/src/api/generated/current-openapi.yaml
openapi: 3.1.0
info:
  title: ShuttleCube Internal Management API
  version: 0.1.0
  description: |
    羽毛球培训与场地经营管理系统的同源内部接口契约。
    该接口仅供已登录管理界面使用，不是外部公共 API。
    `/agents`、`/workflows`、`/agent-runs` 和 `/approval-requests` 为未来 Agent 契约草案，
    当前版本不得实现、生成客户端调用或作为验收范围。
servers:
  - url: /api/v1
security:
  - sessionCookie: []
...
paths:
...
  /classes:
    get:
      tags:
        - Classes
      summary: 查询固定班
      operationId: listClasses
      parameters:
        - name: business_date
          in: query
          schema:
            type: string
            format: date
        - name: ending_within_days
          in: query
          description: 只返回最后一节未来课程在所选天数内的有效班级
          schema:
            type: integer
            enum:
              - 7
              - 15
              - 30
      responses:
        "200":
          description: 固定班列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FixedClassAttentionView"
    post:
      tags:
        - Classes
      summary: 创建固定班并生成课程实例
      operationId: createClass
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FixedClassWrite"
      responses:
        "201":
          description: 固定班及课程实例已创建
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FixedClassCreated"
        "409":
          $ref: "#/components/responses/ScheduleConflict"
        "422":
          $ref: "#/components/responses/ValidationError"
  /classes/{class_id}:
    get:
      tags:
        - Classes
      summary: 查询固定班课程、报名与收费详情
      operationId: getClass
      parameters:
        - name: class_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: 固定班详情，报名收费按实际资金事实汇总
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FixedClassDetail"
        "404":
          description: 固定班不存在
  /classes/{class_id}/capacity:
    patch:
      tags:
        - Classes
      summary: 修改固定班最大人数
      operationId: updateFixedClassCapacity
      parameters:
        - name: class_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FixedClassCapacityWrite"
      responses:
        "200":
          description: 班级容量已更新
        "409":
          $ref: "#/components/responses/ConcurrentChange"
        "422":
          $ref: "#/components/responses/ValidationError"
  /classes/{class_id}/renew:
    post:
      tags:
        - Classes
      summary: 固定班续期并可同步学员课时与应收
      operationId: renewFixedClass
      parameters:
        - name: class_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FixedClassRenewalWrite"
      responses:
        "200":
          description: 新课程及所选学员续期已统一保存
        "409":
          $ref: "#/components/responses/ScheduleConflict"
        "422":
          $ref: "#/components/responses/ValidationError"
  /classes/{class_id}/archive:
    post:
      tags:
        - Classes
      summary: 归档固定班并释放未来排期
      operationId: archiveFixedClass
      parameters:
        - name: class_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FixedClassArchiveWrite"
      responses:
        "200":
          description: 固定班已归档，关联权益已失效
        "409":
          $ref: "#/components/responses/ConcurrentChange"
  /classes/{class_id}/enrollments:
    post:
      tags:
        - Classes
      summary: 为系统内有效学员创建固定班报名（兼容接口）
      description: 管理端日常流程应从学员档案绑定固定班培训权益；该接口保留给既有客户端，并拒绝不存在或已停用的学员。
      operationId: createEnrollment
      parameters:
        - name: class_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EnrollmentWrite"
      responses:
        "201":
          description: 报名及初始课时流水已创建
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Enrollment"
        "404":
          description: 学员或固定班不存在
        "422":
          $ref: "#/components/responses/ValidationError"
  /class-sessions/{session_id}/attendance:finalize:
    post:
      tags:
        - Classes
      summary: 完成课程考勤、扣课并生成教练费用
      operationId: finalizeAttendance
      parameters:
        - name: session_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/IfMatchVersion"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttendanceFinalizeCommand"
      responses:
        "200":
          description: 考勤已完成，返回课时和教练费用结果
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceFinalizeResult"
        "409":
          $ref: "#/components/responses/ConcurrentChange"
        "422":
          $ref: "#/components/responses/ValidationError"
  /class-sessions/{session_id}/cancel-and-replace:
    post:
      tags:
        - Classes
      summary: 机构取消课程并补排
      operationId: cancelAndReplaceClassSession
      parameters:
        - name: session_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/IfMatchVersion"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CancelAndReplaceCommand"
      responses:
        "200":
          description: 原课已取消并创建补排课
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CancelAndReplaceResult"
        "409":
          $ref: "#/components/responses/ScheduleConflict"
        "422":
          $ref: "#/components/responses/ValidationError"
  /class-sessions/{session_id}/reschedule:
    post:
      tags:
        - Classes
      summary: 修改单节固定班课程时间
      operationId: rescheduleClassSession
      parameters:
        - name: session_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClassSessionRescheduleWrite"
      responses:
        "200":
          description: 课程时间及统一排期已更新
        "409":
          $ref: "#/components/responses/ScheduleConflict"
        "422":
          $ref: "#/components/responses/ValidationError"
  /class-sessions/{session_id}/replacement:
    post:
      tags:
        - Classes
      summary: 为待补排的取消课程安排整班补排
      operationId: scheduleCancelledClassSessionReplacement
      parameters:
        - name: session_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClassSessionReplacementWrite"
      responses:
        "201":
          description: 补排课程已生成到课程计划
        "409":
          $ref: "#/components/responses/ScheduleConflict"
        "422":
          $ref: "#/components/responses/ValidationError"
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: shuttlecube_session
  parameters:
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 8
        maxLength: 200
    IfMatchVersion:
      name: If-Match
      in: header
      required: true
      description: 客户端最后读取的实体版本，例如 `"7"`
      schema:
        type: string
...
  responses:
...
    ValidationError:
      description: 输入或业务规则校验失败
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ValidationProblem"
    ConcurrentChange:
      description: 记录版本已变化或目标事实已被其他操作处理
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    ScheduleConflict:
      description: 一个或多个资源发生排期冲突
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ScheduleConflictProblem"
...
  schemas:
...
    FixedClassWrite:
      type: object
      required:
        - name
        - start_date
        - default_start_time
        - duration_minutes
        - session_count
        - capacity
        - default_coach_id
        - court_ids
        - required_court_count
        - student_unit_price
      properties:
        name:
          type: string
          minLength: 1
        class_type:
          type: string
        recurrence_rule:
          type: string
          description: 受支持的每周课程规则
        start_date:
          type: string
          format: date
        default_start_time:
          type: string
          format: time
          description: 必须为整点
        duration_minutes:
          type: integer
          minimum: 60
          multipleOf: 60
        session_count:
          type: integer
          minimum: 1
        capacity:
          type: integer
          minimum: 1
        default_coach_id:
          type: string
          format: uuid
        court_ids:
          type: array
          minItems: 1
          items:
            type: string
            format: uuid
        required_court_count:
          type: integer
          minimum: 1
        student_unit_price:
          $ref: "#/components/schemas/Money"
        coach_fee_per_session:
          anyOf:
            - $ref: "#/components/schemas/Money"
            - type: "null"
          description: 为空时按开班日期采用教练当日标准；保存后冻结为班级单节费用快照
        notes:
          type:
            - string
            - "null"
...
    FixedClassCapacityWrite:
      type: object
      required:
        - capacity
        - version
      properties:
        capacity:
          type: integer
          minimum: 1
        version:
          type: integer
          minimum: 1
...
    FixedClassRenewalWrite:
      type: object
      required:
        - additional_sessions
        - enrollment_renewals
        - version
      properties:
        additional_sessions:
          type: integer
          minimum: 1
        enrollment_renewals:
          type: array
          items:
            $ref: "#/components/schemas/EnrollmentRenewalWrite"
        version:
          type: integer
          minimum: 1
    FixedClassArchiveWrite:
      type: object
      required:
        - reason
        - version
      properties:
        reason:
          type: string
          minLength: 1
          maxLength: 500
        version:
          type: integer
          minimum: 1
    FixedClassAttentionView:
      allOf:
        - $ref: "#/components/schemas/FixedClass"
        - type: object
          properties:
            last_scheduled_end:
              type:
                - string
                - "null"
              format: date-time
            remaining_scheduled_sessions:
              type:
                - integer
                - "null"
              minimum: 0
    FixedClassCreated:
      type: object
      required:
        - class_record
        - session_count
      properties:
        class_record:
          $ref: "#/components/schemas/FixedClass"
        session_count:
          type: integer
        warnings:
          type: array
          items:
            type: string
...
    EnrollmentWrite:
      type: object
      required:
        - student_id
        - enrolled_on
      properties:
        student_id:
          type: string
          format: uuid
        enrolled_on:
          type: string
          format: date
        purchased_units:
          type:
            - integer
            - "null"
          minimum: 1
        actual_receivable:
          anyOf:
            - $ref: "#/components/schemas/Money"
            - type: "null"
        adjustment_reason:
          type:
            - string
            - "null"
    Enrollment:
      allOf:
        - $ref: "#/components/schemas/IdentifierVersion"
        - type: object
          required:
            - suggested_receivable
            - actual_receivable
            - purchased_units
            - status
          properties:
            suggested_receivable:
              $ref: "#/components/schemas/Money"
            actual_receivable:
              $ref: "#/components/schemas/Money"
            purchased_units:
              type: integer
            status:
              type: string
...
    CancelAndReplaceResult:
      type: object
      required:
        - cancelled_session_id
      properties:
        cancelled_session_id:
          type: string
          format: uuid
        replacement_session_id:
          type:
            - string
            - "null"
          format: uuid
...